Create unit test for function that updates the LocationHealthQueue with a new Arrival (or Import)
I tried placing this unit test class in a separate directory inside the Tests directory
AppBundle/Tests/Util, but when running phpunit it doesn't see this class.
So I was forced to place it in the AppBundle/Tests/Controller directory where it does work.
If someone finds a way to make phpunit run it from another directory,
just move this unit test to that more appropriate location.