Add null check for MaediVisna and Scrapie gotten from both the locationHealth and locationHealthMessage The background Likely after testing with some dump and dies in the Arrival and Import logic, DeclareArrivals and DeclareImports were persisted without any locationHealthMessage. This means those DeclareArrivals and DeclareImports would likely not result in any update Health status for the location. And even if the Health of location was updated, ALL of them would be missing the historical records of that mutation because those records are all saved in their respective locationHealthMessages.
The problem
DeclareArrivals and DeclareImports with just a lot of null values in their
locationHealthMessage. This would break the LocationHealth calculation logic.
A simple null check is not enough, because the null values cannot be replaced
by standard default values, if there are still some values present before
the null values.

The fix
Implement nested null checks where the correct replacement values are chosen.