Set weight isActive = false when revoking declareWeight
Note at the moment the isRevoked variable in the weight table is used
instead of the isActive in the measurement table.
This just helps us transition easier by keeping the data consistent
in case we wish to update any logic in the api to use the isActive exclusively
before we remove the isRevoked variable.


Former-commit-id: 479e0c44e4e38db51b37730365e55b84ffc02854